Refactored Block and Transaction
Moved Block and Transaction classes to a new file called xmrtype. This should make code more organized.
